Agregao areaCubierta y cercaniaARival

+++ codigo/Tablero.hpp
@@ -39,7 +39,7 @@ public:
     vector<unsigned int> distJugadorAlArco(const bool enDerecha) const;
     unsigned int distPelotaArco(const bool enDerecha) const;
     unsigned int cercaniaARival(const Jugador &j) const;
-    unsigned int areaCubierta(const bool enDerecha) const;
+    float areaCubierta(const bool enDerecha) const;
 
 private:
